Changed the Transactions files and added class Account

+++ src/main/java/com/github/java_learning/Transactions/Transaction.java
@@ -1,5 +1,7 @@
 package com.github.java_learning.Transactions;
 
 public abstract class Transaction {
-    public abstract int diff();
+
+    public abstract double diff();
+
 }
